split curves.rs into g1.rs, g1p.rs, g2.rs and move push functions to respective structs
