+++ bitvm/src/bn254/mod.rs
@@ -3,8 +3,9 @@ pub mod fq12;
 pub mod fq2;
 pub mod fq6;
 pub mod fr;
-
-pub mod curves;
+pub mod g1;
+pub mod g2;
+pub mod g1p;
 
 pub mod pairing;
 
@@ -13,4 +14,3 @@ pub mod ell_coeffs;
 pub mod fp254impl;
 pub mod msm;
 pub mod utils;
-pub mod g2_subgroup_check;
