+++ bitvm/src/chunker/chunk_g1_points.rs
@@ -118,7 +118,7 @@ mod test {
             }
 
             { msm.clone() }
-            // { g1_affine_push_not_montgomery(expect) }
+            // { G1Affine::push_not_montgomery(expect) }
             // { G1Affine::equalverify() }
             // OP_TRUE
         };
