+++ bitvm/src/groth16/offchain_checker.rs
@@ -182,9 +182,9 @@ mod test {
     use super::*;
     use crate::bn254::ell_coeffs::G2Prepared;
     use crate::bn254::fq12::Fq12;
+    use crate::bn254::fq2::Fq2;
     use crate::bn254::pairing::Pairing;
     use crate::bn254::utils;
-    use crate::bn254::utils::fq12_push;
     use crate::groth16::constants::{LAMBDA, P_POW3};
     use crate::{execute_script_without_stack_limit, treepp::*};
     use ark_bn254::Bn254;
@@ -271,20 +271,20 @@ mod test {
             { utils::from_eval_point(P4.neg()) }
 
             // q4
-            { utils::fq2_push(Q4.x) }
-            { utils::fq2_push(Q4.y) }
+            { Fq2::push(Q4.x) }
+            { Fq2::push(Q4.y) }
 
             // c, c_inv, wi
-            { fq12_push(c) }
-            { fq12_push(c_inv) }
-            { fq12_push(wi) }
+            { Fq12::push(c) }
+            { Fq12::push(c_inv) }
+            { Fq12::push(wi) }
 
             // t4
-            { utils::fq2_push(T4.x) }
-            { utils::fq2_push(T4.y) }
+            { Fq2::push(T4.x) }
+            { Fq2::push(T4.y) }
 
             { quad_miller_loop_with_c_wi.clone() }
-            { fq12_push(hint) }
+            { Fq12::push(hint) }
             { Fq12::equalverify() }
             OP_TRUE
         };
